Conflicts:
	src/com/lds/game/entity/Entity.java
	src/com/lds/physics/CollisionDetector.java

+++ src/com/lds/trigger/EffectRemoveEntity.java
@@ -16,6 +16,16 @@ public class EffectRemoveEntity extends Effect
 	@Override
 	public void fireOutput()
 	{
-		EntityManager.removeEntity(ent);
+	    if (ent != null)
+	    {
+	        EntityManager.removeEntity(ent);
+	        ent = null;
+	    }
+	}
+	
+	@Override
+	public void unfireOutput()
+	{
+	    
 	}
 }
